Move config hash tests to service_test.py
